Use official near-primitives types for receipts
- Add near-primitives 0.34.3 dependency matching near-jsonrpc-client
- Replace custom ReceiptInfo struct with official ReceiptView
- Store full receipt JSON in raw_data instead of partial fields
- Update API to return receipt_id and transaction_hashes arrays
- Simplify receipt handling by using native NEAR types
- Remove unused transaction hash matching logic
- All tests passing
