+++ nt-be/Cargo.toml
@@ -34,6 +34,7 @@ uuid = { version = "1.11", features = ["v4", "serde"] }
 env_logger = "0.11"
 log = "0.4"
 near-jsonrpc-client = "0.20.0"
+near-primitives = "0.34.3"
 
 [dev-dependencies]
 tower = { version = "0.5", features = ["util"] }
